backend-advanced/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
itsdangerous==2.2.0
google-auth==2.30.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.141.0
